requirements.txt
streamlit
chromadb
python-decouple
langchain-openai
langchain
pydantic